FUND THIS PROJECT
I would like to put this website out in order to help the fashion models in My country.
This website is at the stage where it needs editors, more pages and hosting/domain. Your fund will push me to create the difference I want to see for my country.
I have been in the industry that this website is for, 4 years straight and this would be my turning point.
Thank you for your commitment.